sys-libs/pam: Bump and migrate from coreos-overlay to portage-stable
We can now install files to /etc in the normal manner and let the
overlayfs take care of it. We previously relied on tmpfiles to create an
empty /etc/pam.d directory, so get baselayout to do that now.

The pam/libcap dependency loop is not an issue since EAPI 8.

We still apply the locked accounts patch as before.

Closes: https://github.com/flatcar/Flatcar/issues/1349
Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>